feat: fixing news and post feed widget
