+++ lib/widgets/post_feed/post_feed.dart
@@ -1,0 +1,114 @@
+import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+
+import '../../models/post_feed_model.dart';
+import '../../utils/color/color.dart';
+import '../../views/screen/post_feed/post_detail_screen.dart';
+
+class PostFeed extends StatelessWidget {
+  const PostFeed({
+    super.key,
+    required this.postFeeds,
+  });
+
+  final List<PostFeedModel> postFeeds;
+
+  @override
+  Widget build(BuildContext context) {
+    PostFeedModel postFeed = postFeeds[0];
+    return GestureDetector(
+      onTap: () {
+        Navigator.push(
+          context,
+          MaterialPageRoute(
+            builder: (context) => PostDetailScreen(postFeed: postFeed),
+          ),
+        );
+      },
+      child: Card(
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(10),
+        ),
+        child: Padding(
+          padding: const EdgeInsets.all(16),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Row(
+                children: [
+                  const CircleAvatar(
+                    backgroundColor: searchBarTextColor,
+                    child: Icon(
+                      Icons.person,
+                      color: whiteColor,
+                    ),
+                  ),
+                  const SizedBox(width: 8),
+                  Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Text(
+                        postFeed.name ?? '',
+                        style: GoogleFonts.poppins(
+                          fontWeight: FontWeight.w600,
+                          fontSize: 12,
+                        ),
+                      ),
+                      const SizedBox(height: 4),
+                      Text(
+                        postFeed.time ?? '',
+                        style: GoogleFonts.poppins(
+                          fontWeight: FontWeight.w400,
+                          fontSize: 10,
+                        ),
+                      ),
+                    ],
+                  ),
+                ],
+              ),
+              const SizedBox(height: 8),
+              Text(
+                postFeed.comment ?? '',
+                style: GoogleFonts.poppins(
+                  fontWeight: FontWeight.w400,
+                  fontSize: 14,
+                ),
+              ),
+              const SizedBox(height: 10),
+              Row(
+                children: [
+                  IconButton(
+                    icon: const Icon(
+                      Icons.thumb_up,
+                      color: Colors.blue,
+                    ),
+                    onPressed: () {},
+                  ),
+                  Text(
+                    postFeed.like ?? '',
+                    style: GoogleFonts.poppins(
+                      fontWeight: FontWeight.w400,
+                      fontSize: 11,
+                    ),
+                  ),
+                  const SizedBox(width: 15),
+                  IconButton(
+                    icon: const Icon(Icons.comment_outlined),
+                    onPressed: () {},
+                  ),
+                  Text(
+                    postFeed.reply ?? '',
+                    style: GoogleFonts.poppins(
+                      fontWeight: FontWeight.w400,
+                      fontSize: 11,
+                    ),
+                  ),
+                ],
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
